Сервис-альтернатива MessageBox

+++ Tests/Yasinovsky.MailSender.Tests.Wpf/MainWindow.xaml.cs
@@ -14,56 +14,64 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using Yasinovsky.MailSender.Core.Extensions;
 using Yasinovsky.MailSender.Core.Contracts.Services;
+using Yasinovsky.MailSender.Core.Enums;
 using Yasinovsky.MailSender.Services;
+using Yasinovsky.MailSender.Services.Wpf;
 
 namespace Yasinovsky.MailSender.Tests.Wpf
 {
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : Window
+    public partial class MainWindow
     {
-        private INetUserService _netUserService;
-        private ISmtpClientAccessor _smtpClientAccessor;
-        private IEmailSendService _emailSendService;
-
+        private readonly INetUserService _netUserService;
+        private readonly ISmtpClientAccessor _smtpClientAccessor;
+        private readonly IEmailSendService _emailSendService;
+        private readonly IUserDialogService _userDialogService;
         public MainWindow()
         {
             InitializeComponent();
             _netUserService = new NetUserService();
             _smtpClientAccessor = new SmtpClientAccessor(_netUserService);
             _emailSendService = new SmtpEmailSendService(_smtpClientAccessor);
+            _userDialogService = new CustomWindowUserDialogService();
         }
 
         private async void SendButton_Click(object sender, RoutedEventArgs e)
         {
-            IsEnabled = false;
-            if ((bool)IsAuthenticateCheck.IsChecked)
-            {
-                _netUserService.SetCredentials(EmailEdit.Text, PasswordEdit.SecurePassword);
-            }
-            else
-            {
-                _netUserService.RemoveCurrentCredentials();
-            }
-            if (!int.TryParse(PortEdit.Text, out int port))
-                return;
-            _smtpClientAccessor.SetClient(
-                HostEdit.Text,
-                port,
-                (bool)IsSslCheck.IsChecked);
             try
             {
+                IsEnabled = false;
+                if (IsAuthenticateCheck.IsChecked != null && (bool) IsAuthenticateCheck.IsChecked)
+                    _netUserService.SetCredentials(EmailEdit.Text, PasswordEdit.SecurePassword);
+                else
+                    _netUserService.RemoveCurrentCredentials();
+                if (!int.TryParse(PortEdit.Text, out var port))
+                    return;
+                _smtpClientAccessor.SetClient(
+                    HostEdit.Text,
+                port,
+                IsSslCheck.IsChecked != null && (bool)IsSslCheck.IsChecked);
+            
                 var result = await _emailSendService.SendAsync(
                     SubjectEdit.Text,
                     BodyEdit.Text,
                     EmailEdit.Text,
                     EmailToEdit.Text);
-                MessageBox.Show(result.ToString());
+                if (result == EmailSendResult.Success)
+                {
+                    await _userDialogService.ShowInformationAsync("Email send success", "Info");
+                }
+                else if (result == EmailSendResult.Unauthorized)
+                {
+                    await _userDialogService.ShowWarningAsync("Unauthorized", "Warning");
+                }
+                else
+                {
+                    await _userDialogService.ShowErrorAsync("Timeout", "Error");
+                }
             }
             catch (Exception exp)
             {
-                MessageBox.Show(exp.Message, exp.GetType().FullName);
+                await _userDialogService.ShowErrorAsync($"{exp.Message}\n{exp.StackTrace}", exp.GetType().FullName);
             }
             IsEnabled = true;
         }
